Develop

+++ src/postgres_database/database.py
@@ -1,9 +1,12 @@
 import psycopg2
 from config import config
+from src.utils import Logger
+import inspect
 
 class Database:
     def __init__(self):
         params = config()
+        self.logger = Logger(__name__)
         self.connection = psycopg2.connect(**params)
         self.cursor = self.connection.cursor()
 
@@ -12,25 +15,25 @@ class Database:
             values = ', '.join(columns) if isinstance(columns, list) else columns if ', ' in columns else columns.replace(' ', ', ')
             command = f'SELECT {values} FROM {table};' if conditions == '' else f'SELECT {values} FROM {table} WHERE {conditions};'
             self.cursor.execute(command)
-        except psycopg2.errors.UndefinedColumn as err:
-            raise(err)
-        except psycopg2.errors.InFailedSqlTransaction as err:
-            raise(err) 
-        return self.cursor.fetchall()
+            return self.cursor.fetchall()
+        except Exception as e:
+            self.connection.rollback()
+            self.logger.logger.error(f"An error occurred: {e}")
     
     def insert(self, table, values):
         try:
             command_table_part = f"INSERT INTO {table}{tuple(self.get_table_columns_names(table))}".replace("'", "")
             command_value_part = f"VALUES{tuple(values)} RETURNING test_ident;"
             self.cursor.execute(f'{command_table_part} {command_value_part}');
-        except psycopg2.errors.DatabaseError as err:
-            raise(err)
-        except psycopg2.errors.SyntaxError as err:
-            raise(err)
-        return self.cursor.fetchone()[0]
+            return self.cursor.fetchone()[0]
+        except Exception as e:
+            self.connection.rollback()
+            self.logger.logger.error(f"An error occurred: {e}")
     
-    def update(self, table, columns, new_values, ident_column, ident_value):
+    def update(self, table=None, columns=None, new_values=None, ident_column=None, ident_value=None):
         try:
+            if any(value is None for value in locals().values()):
+                raise Exception('One of parameters is None!')    
             command_set_new_values = ''
             if isinstance(columns, list) and isinstance(new_values, list):
                 for i in range(len(columns)):
@@ -38,22 +41,23 @@ class Database:
                     command_set_new_values += ',' if i < len(columns)-1 else ''
                 command_returning = f'RETURNING {columns[0]};'.replace("'", "")
             else:
-                raise(Exception("Column and new value have to be the same type!"))
+                raise Exception("Column and new value have to be the same type!")
             self.cursor.execute(f'UPDATE {table} SET{command_set_new_values} WHERE {ident_column} = {ident_value} {command_returning}')   
-        except IndexError as err:
-            raise(err)
-        except TypeError as err:
-            raise(err)
-        except psycopg2.errors.SyntaxError as err:
-            raise(err)
-        return self.cursor.fetchone()[0]
+            return self.cursor.fetchone()[0]
+        except Exception as e:
+            self.connection.rollback()
+            self.logger.logger.error(f"An error occurred: {e}")
+
     
-    def delete(self, table, condition):
+    def delete(self, table=None, condition=None):
         try:
+            if any(value is None for value in locals().values()):
+                raise Exception('One of parameters is None!') 
             self.cursor.execute(f'DELETE FROM {table} WHERE {condition} RETURNING id;')
-        except psycopg2.errors.SyntaxError as err:
-            raise(err)
-        return self.cursor.fetchone()[0]
+            return self.cursor.fetchone()[0]
+        except Exception as e:
+            self.connection.rollback()
+            self.logger.logger.error(f"An error occurred: {e}")
         
 
     def get_table_columns_names(self, table):
